Report test result by gdb exit code.

+++ run_tests.py
@@ -21,49 +21,7 @@ def safe_decode(data):
     return s.replace('\r', '')
 
 
-def send_command(gdb, cmd):
-    try:
-        gdb.expect_exact(['(gdb)'], timeout=3)
-    except pexpect.exceptions.TIMEOUT:
-        gdb.kill(signal.SIGINT)
-    gdb.sendline(cmd)
-    print(safe_decode(gdb.before), end='', flush=True)
-    print(safe_decode(gdb.after), end='', flush=True)
-
-
-# Tries to start gdb in order to investigate kernel state on deadlock or crash.
-def gdb_inspect(interactive):
-    gdb_cmd = getvar('gdb.binary')
-    gdb_port = getvar('config.gdbport')
-    if interactive:
-        gdb_opts = ['-iex=set auto-load safe-path {}/'.format(os.getcwd()),
-                    '-ex=target remote localhost:%u' % gdb_port,
-                    '--silent', getvar('config.kernel')]
-    else:
-        # Note: These options are different than .gdbinit.
-        gdb_opts = ['-ex=target remote localhost:%u' % gdb_port,
-                    '-ex=python import os, sys',
-                    '-ex=python sys.path.append(os.getcwd() + "/sys")',
-                    '-ex=python import debug',
-                    '-ex=set pagination off',
-                    '--nh', '--nx', '--silent', getvar('config.kernel')]
-    gdb = pexpect.spawn(gdb_cmd, gdb_opts, timeout=3)
-    if interactive:
-        send_command(gdb, 'backtrace full')
-        send_command(gdb, 'kthread')
-        gdb.interact()
-    else:
-        send_command(gdb, 'info registers')
-        send_command(gdb, 'cpu tlb')
-        send_command(gdb, 'backtrace full')
-        # following commands may fail
-        send_command(gdb, 'kthread')
-        send_command(gdb, 'klog')
-        # we need dummy command - otherwise previous one won't appear
-        send_command(gdb, 'quit')
-
-
-def test_seed(seed, interactive=True, repeat=1, retry=0):
+def test_seed(seed, repeat=1, retry=0):
     if retry == RETRIES_MAX:
         print("Maximum retries reached, still not output received. "
               "Test inconclusive.")
@@ -75,14 +33,12 @@ def test_seed(seed, interactive=True, repeat=1, retry=0):
                            '--port', str(getvar('config.gdbport')),
                            '-t', 'test=all', 'klog-quiet=1',
                            'seed=%u' % seed, 'repeat=%d' % repeat])
-    index = child.expect_exact(
-        ['kprintf("Test run finished!\\n");', 'panic_fail(void)', pexpect.EOF,
-            pexpect.TIMEOUT],
-        timeout=TIMEOUT)
+    index = child.expect_exact([pexpect.EOF, pexpect.TIMEOUT], timeout=TIMEOUT)
     if index == 0:
+        child.close()
+        if child.exitstatus == 0:
+            return
         child.terminate(True)
-        return
-    elif index == 1:
         print("Test failure reported!\n")
         message = safe_decode(child.before)
         message += safe_decode(child.buffer)
@@ -92,27 +48,17 @@ def test_seed(seed, interactive=True, repeat=1, retry=0):
         except pexpect.exceptions.TIMEOUT:
             pass
         print(message)
-        gdb_inspect(interactive)
         sys.exit(1)
-    elif index == 2:
-        message = safe_decode(child.before)
-        message += safe_decode(child.buffer)
-        print(message)
-        print("EOF reached without success report. This may indicate "
-              "a problem with the testing framework or QEMU. "
-              "Retrying (%d)..." % (retry + 1))
-        test_seed(seed, interactive, repeat, retry + 1)
-    elif index == 3:
+    elif index == 1:
         print("Timeout reached!\n")
         message = safe_decode(child.buffer)
+        child.terminate(True)
         print(message)
         if len(message) < 100:
             print("It looks like kernel did not even start within the time "
                   "limit. Retrying (%d)..." % (retry + 1))
-            child.terminate(True)
-            test_seed(seed, interactive, repeat, retry + 1)
+            test_seed(seed, repeat, retry + 1)
         else:
-            gdb_inspect(interactive)
             print("No test result reported within timeout. Unable to verify "
                   "test success. Seed was: %u, repeat: %d" % (seed, repeat))
             sys.exit(1)
@@ -131,8 +77,6 @@ if __name__ == '__main__':
         description='Automatically performs kernel tests.')
     parser.add_argument('--times', type=int, default=N_SIMPLE,
                         help='Run tests given number of times.')
-    parser.add_argument('--non-interactive', action='store_true',
-                        help='Do not run gdb session if tests fail.')
     parser.add_argument('--board', default='malta', choices=['malta', 'rpi3'],
                         help='Emulated board.')
     args = parser.parse_args()
@@ -140,12 +84,10 @@ if __name__ == '__main__':
     setboard(args.board)
     setvar('config.gdbport', RandomPort())
 
-    interactive = not args.non_interactive
-
     # Run tests using n random seeds
     for i in range(0, args.times):
         seed = random.randint(0, 2**32)
-        test_seed(seed, interactive, REPEAT)
+        test_seed(seed, REPEAT)
 
     print("Tests successful!")
     sys.exit(0)
